telegram_bot/requirements.txt
telethon==1.40.0
pymongo==4.13.0
openai>=1.0.0
anthropic>=0.5.0
mistralai>=0.0.7
python-i18n==0.3.9
python-dotenv==1.0.0
pillow>=10.0.0
requests>=2.0.0